🔧 Refactor code for world and item creation
This commit refactors the code for creating worlds and items. It adds two new collections, "world-1" and "world-2", to the environment keys. The code also creates a red party hat world token and a blue party hat world token using the new collection policy. Finally, it updates the gielinorian-item-policy.pact file with version 2 of the policy.